feature: Extracts cross-concern helpers into one internal kernel
